move writeMediaSection to separate file
